moved application to Posttype folder
