+++ source/php/PostType/Application/ApplicationConfiguration.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace VolunteerManager;
+namespace VolunteerManager\PostType\Application;
 
 use VolunteerManager\Helper\Icon as Icon;
